Fix waitpid() and kill() behavior

+++ autoload/proc.c
@@ -759,10 +759,12 @@ vp_kill(char *args)
         return vp_stack_return_error(&_result, "kill() error: %s",
                 strerror(errno));
 
-    /* Kill by the process group. */
-    pgid = getpgid(pid);
-    if (pgid > 0) {
-        kill(-pgid, sig);
+    if (sig != 0) {
+        /* Kill by the process group. */
+        pgid = getpgid(pid);
+        if (pgid > 0) {
+            kill(-pgid, sig);
+        }
     }
 
     vp_stack_push_num(&_result, "%d", ret);
@@ -785,15 +787,15 @@ vp_waitpid(char *args)
         return vp_stack_return_error(&_result, "waitpid() error: %s",
                 strerror(errno));
     if (n == 0 || WIFCONTINUED(status)) {
+        vp_stack_push_str(&_result, "run");
+        vp_stack_push_num(&_result, "%d", 0);
+    } else if (WIFEXITED(status)) {
         /* Kill by the process group. */
         pgid = getpgid(pid);
         if (pgid > 0) {
             kill(-pgid, 15);
         }
 
-        vp_stack_push_str(&_result, "run");
-        vp_stack_push_num(&_result, "%d", 0);
-    } else if (WIFEXITED(status)) {
         vp_stack_push_str(&_result, "exit");
         vp_stack_push_num(&_result, "%d", WEXITSTATUS(status));
     } else if (WIFSIGNALED(status)) {
